+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 * Deprecated the experimental configuration function enableServerConfiguration. It is now enabled by default and can be controlled directly from the server.
 
+* Removed the deprecated function "Countly.apm.triggerForeground()"
+* Removed the deprecated function "Countly.apm.triggerBackground()"
+
 ## 25.1.1
 * Mitigated an issue where after closing a content, they were not being fetched again.